chapter_15/section_02/rag/requirements.txt
openai
python-dotenv
langchain>=0.0.331
langchain-openai
langchain-core
numpy

deepeval
chromadb 
langchain_community 
langchain_text_splitters
